Asyncify version of Clangd uploaded, and Clangd.mjs modified.
It is now possible to cause a wait to the clangd WASM module, using asynctify, this can be done at in out stdstream class, due to the addition of async alternative methods in calngd.mjs. Current implementation just waits 4ms (this can probably work as wait indefinitely until x happens), but it does mean LSP requests can be qued up as the thread is no longer blocked whilst it tries to read the stdio. Currently only send a request every 4 characters typed. Maybe set up a flag that returns the enter key on first encounter with a zero, then continue with the usual wait.