Export a generic fs interface for filesystems to be defined from
Also disabled some tests that failed due to these changes. This should be reenabled in a hex gen test file in the future.

Simulator was modified for hex gen rework to address on oversight from the previous commit, however the simulator is set to be removed anyway so this change isn't especially relevant.