Generalised the fs interface name and its method names
Renamed some python specific naming to be more general since python is slowly being removed.

Also included additional methods in the interface that were being called in the UI.

New filesystems should be able to be written now. Going forward, we will write a simple filesystem that does not implement the python specific functionality (namely MicropythonFS) and is catered for the C++ project instead.

The original filesystem has been left in place for now, renamed as _FileSystem