remove api docs lsp request
This feature caused clangd to crash much quicker. In fact 2 maximum stack size errors are reported in the console when this feature is enabled.

Since api docs population is hacky anyway, it might be worth just leaving this out for now. Might make it just a little bit easier to fix #10